More package refactoring
